requirements.txt
inferactively-pymdp
numpy
matplotlib